requirements.txt
boto3==1.24.24
moviepy==1.0.3
mutagen==1.46.0
Pillow==9.4.0
playsound==1.3.0
playwright==1.23.0
praw==7.6.1
python_vlc==3.0.18121
pytube==12.1.0
requests==2.28.1
rich==13.3.1
